+++ Data/SBV/Tools/Polynomial.hs
@@ -134,7 +134,7 @@ ites s xs ys
 
 -- | Multiply two polynomials and reduce by the third (concrete) irreducible, given by its coefficients.
 -- See the remarks for the 'pMult' function for this design choice
-polyMult :: (Num a, Bits a, SymWord a, FromBits (SBV a)) => (SBV a, SBV a, [Int]) -> SBV a
+polyMult :: (SFiniteBits a, FromBits (SBV a)) => (SBV a, SBV a, [Int]) -> SBV a
 polyMult (x, y, red)
   | isReal x
   = error $ "SBV.polyMult: Received a real value: " ++ show x
@@ -149,7 +149,7 @@ polyMult (x, y, red)
         mul _  []     ps = ps
         mul as (b:bs) ps = mul (false:as) bs (ites b (as `addPoly` ps) ps)
 
-polyDivMod :: (Num a, Bits a, SymWord a, FromBits (SBV a)) => SBV a -> SBV a -> (SBV a, SBV a)
+polyDivMod :: (SFiniteBits a, FromBits (SBV a)) => SBV a -> SBV a -> (SBV a, SBV a)
 polyDivMod x y
    | isReal x
    = error $ "SBV.polyDivMod: Received a real value: " ++ show x
@@ -240,7 +240,7 @@ crcBV n m p = take n $ go (replicate n false) (m ++ replicate n false)
 
 -- | Compute CRC's over polynomials, i.e., symbolic words. The first
 -- 'Int' argument plays the same role as the one in the 'crcBV' function.
-crc :: (FromBits (SBV a), FromBits (SBV b), Num a, Num b, Bits a, Bits b, SymWord a, SymWord b) => Int -> SBV a -> SBV b -> SBV b
+crc :: (FromBits (SBV a), FromBits (SBV b), SFiniteBits a, SFiniteBits b) => Int -> SBV a -> SBV b -> SBV b
 crc n m p
   | isReal m || isReal p
   = error $ "SBV.crc: Received a real value: " ++ show (m, p)